src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ./requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.31.0
    # via -r requirements.in
bcrypt==5.0.0
    # via -r requirements.in
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic
fastapi[standard]==0.123.0
    # via
    #   -r requirements.in
    #   fastapi-pagination
    #   fastapi-sqlalchemy-toolkit
fastapi-cli[standard]==0.0.16
    # via fastapi
fastapi-cloud-cli==0.5.2
    # via fastapi-cli
fastapi-pagination==0.15.3
    # via fastapi-sqlalchemy-toolkit
fastapi-sqlalchemy-toolkit==0.7.17
    # via -r requirements.in
fastar==0.8.0
    # via fastapi-cloud-cli
greenlet==3.3.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pycountry==24.6.1
    # via -r requirements.in
pydantic[email]==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cloud-cli
    #   fastapi-pagination
    #   fastapi-sqlalchemy-toolkit
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.10.6
    # via -r requirements.in
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.3
    # via uvicorn
redis==7.1.0
    # via -r requirements.in
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.17.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk==2.46.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sqlalchemy==2.0.44
    # via
    #   -r requirements.in
    #   fastapi-sqlalchemy-toolkit
starlette==0.50.0
    # via fastapi
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   fastapi-pagination
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via sentry-sdk
uvicorn[standard]==0.38.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
